training-requirements.txt
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.0
debugpy==1.8.14
decorator==5.2.1
easydict==1.13
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.17.0
onnxruntime-gpu==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prettytable==3.16.0
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.4.0
simsimd==6.2.1
six==1.17.0
stack-data==0.6.3
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
